+++ bluetoe/sm/include/bluetoe/oob_authentication.hpp
@@ -6,6 +6,7 @@
 #include <cstdint>
 #include <array>
 #include <utility>
+#include <tuple>
 
 namespace bluetoe {
