+++ bluetoe/sm/include/bluetoe/security_manager.hpp
@@ -5,6 +5,7 @@
 #include <cstdint>
 #include <cassert>
 #include <array>
+#include <algorithm>
 
 #include <bluetoe/codes.hpp>
 #include <bluetoe/address.hpp>
